# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Collaborative Editor AI Backend API
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    LoginRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
        displayName:
          type: string
    LoginResponse:
      type: object
      required: [userId, displayName, globalRole, accessToken, expiresIn]
      properties:
        userId:
          type: string
        displayName:
          type: string
        globalRole:
          type: string
          enum: [user, admin]
        accessToken:
          type: string
        expiresIn:
          type: integer
    CreateDocumentRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string
    CreateDocumentResponse:
      type: object
      required: [documentId, title, ownerId, createdAt, updatedAt, currentVersionId]
      properties:
        documentId:
          type: string
        title:
          type: string
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        currentVersionId:
          type: string
    GetDocumentResponse:
      type: object
      required: [documentId, title, content, updatedAt, currentVersionId, role, revisionId]
      properties:
        documentId:
          type: string
        title:
          type: string
        content:
          type: string
        updatedAt:
          type: string
          format: date-time
        currentVersionId:
          type: string
        role:
          type: string
          enum: [owner, editor, viewer]
        revisionId:
          type: string
    UpdateContentRequest:
      type: object
      required: [requestId, content, baseRevisionId]
      properties:
        requestId:
          type: string
        content:
          type: string
        baseRevisionId:
          type: string
    VersionListResponse:
      type: object
      properties:
        documentId:
          type: string
        versions:
          type: array
          items:
            type: object
            properties:
              versionId:
                type: string
              versionNumber:
                type: integer
              createdAt:
                type: string
              createdBy:
                type: string
              reason:
                type: string
        nextCursor:
          type: string
          nullable: true
    PermissionUpdateRequest:
      type: object
      required: [requestId, targetUserId, role]
      properties:
        requestId:
          type: string
        targetUserId:
          type: string
        role:
          type: string
          enum: [viewer, editor]
    SessionRequest:
      type: object
      required: [documentId]
      properties:
        documentId:
          type: string
    AiSelection:
      type: object
      required: [start, end]
      properties:
        start:
          type: integer
        end:
          type: integer
paths:
  /health:
    get:
      responses:
        "200":
          description: OK
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /documents:
    post:
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateDocumentResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{documentId}:
    get:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDocumentResponse"
  /documents/{documentId}/content:
    put:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentRequest"
      responses:
        "200":
          description: Updated
        "409":
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{documentId}/versions:
    get:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Versions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionListResponse"
  /documents/{documentId}/revert:
    post:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
  /documents/{documentId}/permissions:
    get:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
    put:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionUpdateRequest"
  /documents/{documentId}/permissions/{targetUserId}:
    delete:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
        - in: path
          name: targetUserId
          required: true
          schema:
            type: string
  /documents/{documentId}/ai-policy:
    get:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
    put:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
  /sessions:
    post:
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionRequest"
  /ai/rewrite:
    post:
      security:
        - BearerAuth: []
  /ai/summarize:
    post:
      security:
        - BearerAuth: []
  /ai/translate:
    post:
      security:
        - BearerAuth: []
  /ai/jobs/{jobId}:
    get:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
  /documents/{documentId}/export:
    post:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
  /exports/{jobId}:
    get:
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string